--- app/components/Views/SendFlow/AddressFrom/index.test.tsx
@@ -1,1 +1,0 @@
-// todo
