+++ app/constants/navigation/Routes.ts
@@ -45,6 +45,7 @@ const Routes = {
   },
   SEND_FLOW: {
     SEND_TO: 'SendTo',
+    AMOUNT: 'Amount',
   },
   ACCOUNT_BACKUP: {
     STEP_1_B: 'AccountBackupStep1B',
